separated the text style into a separate file
