+++ src/engraving/libmscore/read206.cpp
@@ -3403,6 +3403,7 @@ static void readStyle(MStyle* style, XmlReader& e)
             style->chordList()->clear();
             style->chordList()->read(e);
             style->setCustomChordList(true);
+            //! TODO Look like a bug, changes just the local variable
             for (ChordFont f : style->chordList()->fonts) {
                 if (f.family == "MuseJazz") {
                     f.family = "MuseJazz Text";
